Move buf constructors to pixel
All `unsafe` is now contained in `pixel.rs`. This should make safety
review easier, in particular the witness type Pixel refers to this
constructor for its own soundness.
